+++ src/bb_web_ds_tools/views/malli.cljs
@@ -476,8 +476,7 @@
                [c/label "Schema (EDN)"]
                [c/help-button
                 {:href (nav/get-wiki-url :malli)
-                 :title "Help: Malli Validation"
-                 :class "!p-1 !w-5 !h-5 opacity-50 hover:opacity-100 mb-2"}]]
+                 :title "Help: Malli Validation"}]]
               [c/button {:size :sm
                          :variant :primary
                          :on-click #(rf/dispatch [:malli/validate])} "Validate"]
@@ -505,8 +504,7 @@
                [c/label "Schema (EDN)"]
                [c/help-button
                 {:href (nav/get-wiki-url :malli)
-                 :title "Help: JSON Schema"
-                 :class "!p-1 !w-5 !h-5 opacity-50 hover:opacity-100 mb-2"}]]
+                 :title "Help: JSON Schema"}]]
               [c/button {:size :sm
                          :variant :primary
                          :on-click #(rf/dispatch [:malli/transform-json])} "Transform to JSON Schema"]]